Singleactivity

+++ app/src/main/java/de/prog3/ackerschlagkartei/data/repositories/FirestoreRepository.java
@@ -9,6 +9,7 @@ import androidx.lifecycle.MutableLiveData;
 
 import com.google.android.gms.tasks.OnFailureListener;
 import com.google.android.gms.tasks.OnSuccessListener;
+import com.google.firebase.FirebaseApp;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.firestore.CollectionReference;
 import com.google.firebase.firestore.DocumentReference;
@@ -16,6 +17,7 @@ import com.google.firebase.firestore.DocumentSnapshot;
 import com.google.firebase.firestore.EventListener;
 import com.google.firebase.firestore.FirebaseFirestore;
 import com.google.firebase.firestore.FirebaseFirestoreException;
+import com.google.firebase.firestore.FirebaseFirestoreSettings;
 import com.google.firebase.firestore.Query;
 import com.google.firebase.firestore.QueryDocumentSnapshot;
 import com.google.firebase.firestore.QuerySnapshot;
@@ -51,8 +53,16 @@ public class FirestoreRepository {
 
     public FirestoreRepository(@NonNull Application application) {
         this.application = application;
+
+        // Enable Offline Persistence
+        FirebaseFirestoreSettings settings = new FirebaseFirestoreSettings.Builder()
+                .setPersistenceEnabled(true)
+                .build();
+        FirebaseFirestore.getInstance().setFirestoreSettings(settings);
+
         this.firebaseFirestore = FirebaseFirestore.getInstance();
         this.firebaseAuth = FirebaseAuth.getInstance();
+
 
         this.fieldListGetData = new MutableLiveData<>();
         this.fieldGetData = new MutableLiveData<>();
@@ -72,9 +82,6 @@ public class FirestoreRepository {
     }
 
     private CollectionReference getFieldCollection() {
-        if (this.firebaseAuth.getCurrentUser() == null) {
-            throw new NullPointerException("User is not logged in.");
-        }
 
         String uid = this.firebaseAuth.getUid();
 
@@ -87,6 +94,8 @@ public class FirestoreRepository {
     // FIELDMODEL
 
     public MutableLiveData<List<FieldModel>> getFieldListGetData() {
+        if(firebaseAuth.getCurrentUser() == null) return fieldListGetData;
+
         this.fieldListGetStatus.postValue(Status.LOADING);
         this.getFieldCollection().addSnapshotListener(new EventListener<QuerySnapshot>() {
             @Override
@@ -94,7 +103,6 @@ public class FirestoreRepository {
 
                 if (error != null) {
                     fieldListGetStatus.postValue(Status.ERROR);
-                    Toast.makeText(application, error.getLocalizedMessage(), Toast.LENGTH_SHORT).show();
                     return;
                 }
 
@@ -118,6 +126,8 @@ public class FirestoreRepository {
     }
 
     public MutableLiveData<FieldModel> getFieldMutableLiveData(@NonNull String uid) {
+        if(firebaseAuth.getCurrentUser() == null) return fieldGetData;
+
         this.fieldGetStatus.postValue(Status.LOADING);
         this.getFieldCollection().document(uid).addSnapshotListener(new EventListener<DocumentSnapshot>() {
             @Override
@@ -125,7 +135,6 @@ public class FirestoreRepository {
 
                 if (error != null) {
                     fieldGetStatus.postValue(Status.ERROR);
-                    Toast.makeText(application, error.getLocalizedMessage(), Toast.LENGTH_SHORT).show();
                     return;
                 }
 
@@ -142,6 +151,8 @@ public class FirestoreRepository {
     }
 
     public void createFieldModel(@NonNull FieldModel fieldModel) {
+        if(firebaseAuth.getCurrentUser() == null) return;
+
         this.fieldCreateStatus.postValue(Status.LOADING);
         this.getFieldCollection().add(fieldModel).addOnSuccessListener(new OnSuccessListener<DocumentReference>() {
             @Override
@@ -158,6 +169,8 @@ public class FirestoreRepository {
     }
 
     public void updateFieldModel(@NonNull FieldModel fieldModel, String field, Object changes) {
+        if(firebaseAuth.getCurrentUser() == null) return;
+
         this.fieldUpdateStatus.postValue(Status.LOADING);
         this.getFieldCollection()
                 .document(fieldModel.getUid())
@@ -177,6 +190,8 @@ public class FirestoreRepository {
     }
 
     public void deleteFieldModel(@NonNull FieldModel fieldModel) {
+        if(firebaseAuth.getCurrentUser() == null) return;
+
         this.fieldDeleteStatus.postValue(Status.LOADING);
         this.getFieldCollection().document(fieldModel.getUid()).delete().addOnSuccessListener(new OnSuccessListener<Void>() {
             @Override
@@ -195,6 +210,8 @@ public class FirestoreRepository {
     // ACTIONMODEL
 
     public MutableLiveData<List<ActionModel>> getActionListGetData(@NonNull String fieldUid, String category) {
+        if(firebaseAuth.getCurrentUser() == null) return actionListGetData;
+
         this.actionListGetStatus.postValue(Status.LOADING);
         this.getFieldCollection()
                 .document(fieldUid)
@@ -207,8 +224,6 @@ public class FirestoreRepository {
 
                         if (error != null) {
                             actionListGetStatus.postValue(Status.ERROR);
-                            //Log.w("GetAction", error.getLocalizedMessage());
-                            //Toast.makeText(application, error.getLocalizedMessage(), Toast.LENGTH_SHORT).show();
                             return;
                         }
 
@@ -232,6 +247,8 @@ public class FirestoreRepository {
     }
 
     public void createActionModel(@NonNull String fieldUid, ActionModel actionModel) {
+        if(firebaseAuth.getCurrentUser() == null) return;
+
         this.actionCreateStatus.postValue(Status.LOADING);
         this.getFieldCollection()
                 .document(fieldUid)
@@ -246,7 +263,6 @@ public class FirestoreRepository {
             @Override
             public void onFailure(@NonNull Exception e) {
                 actionCreateStatus.postValue(Status.ERROR);
-                //Toast.makeText(application, e.getLocalizedMessage(), Toast.LENGTH_SHORT).show();
             }
         });
     }
@@ -254,6 +270,8 @@ public class FirestoreRepository {
     // DOCUMENTMODEL
 
     public void createDocumentModel(@NonNull String fieldUid, DocumentModel documentModel) {
+        if(firebaseAuth.getCurrentUser() == null) return;
+
         this.documentCreateStatus.postValue(Status.LOADING);
         this.getFieldCollection()
                 .document(fieldUid)
@@ -268,12 +286,13 @@ public class FirestoreRepository {
             @Override
             public void onFailure(@NonNull Exception e) {
                 documentCreateStatus.postValue(Status.ERROR);
-                //Toast.makeText(application, e.getLocalizedMessage(), Toast.LENGTH_SHORT).show();
             }
         });
     }
 
     public MutableLiveData<List<DocumentModel>> getDocumentListGetData(@NonNull String fieldUid) {
+        if(firebaseAuth.getCurrentUser() == null) return documentListGetData;
+
         this.documentListGetStatus.postValue(Status.LOADING);
         this.getFieldCollection()
                 .document(fieldUid)
@@ -284,7 +303,6 @@ public class FirestoreRepository {
 
                         if (error != null) {
                             documentListGetStatus.postValue(Status.ERROR);
-                            //Toast.makeText(application, error.getLocalizedMessage(), Toast.LENGTH_SHORT).show();
                             return;
                         }
 
@@ -306,7 +324,7 @@ public class FirestoreRepository {
         return documentListGetData;
     }
 
-    // STATUS GETTER
+    // STATUS GETTER FOR OPTIONAL UI UPDATES
 
     public MutableLiveData<Status> getFieldListGetStatus() {
         return fieldListGetStatus;
